+++ demo/pubsub/topology.js
@@ -1,3 +1,5 @@
+'use strict'
+
 module.exports = function( rabbit, subscribeTo ) {
 	return rabbit.configure( {
 		// arguments used to establish a connection to a broker